Make errors very explicit
If you're already invited but haven't signed in, you probably missed the
email from feedback@slack.com. Added an error message about this.

Also, if you're already on the team, you need to go to the signin page.
There's a new link for it but let's output the page in the error message.

Updated styles so multiline errors will display.