Move make from Procfile to postinstall
Heroku's Procfile lists the commands to be run each time
the dyno starts. Dynos are started at minimum once a day, so
the make process would happen daily. By moving it to postinstall,
the built files are part of the slug, which is deployed before the
dyno starts. This should significantly reduce dyno start time.