Dockerfile bug fixes
npm install wasn't actually running - octohost/nodejs was throwing
an error on npm install because it couldn't drop privs in the
workdir. Adding --unsafe-perm seems okay in a docker context.

If there is no SLACK_CHANNELS env var, slackin was failing to parse
the command line. I updated it to match the Heroku command line.