Update docs with more clear token generation info
As a new user of slackin I did not understand that I needed only a test token and that creating a whole OAuth 2 token was not needed and in fact would not work. This update the docs to be very clear that you should be using a test token with Slackin.